New crafting achievement script, fixes to others
Restructuring and bug fixes for achievement_checker.

Restructuring for artifact_locations.

crafting.py is new. This script looks up items owned by the player
(inventory, fridge, chests) and identifies which items that have not
yet been crafted can be, and which items are missing to be able to do
so.